Basic endpoints

+++ src/main/java/com/kazmiruk/travel_agency/repository/TourRepository.java
@@ -3,5 +3,5 @@ package com.kazmiruk.travel_agency.repository;
 import com.kazmiruk.travel_agency.model.Tour;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TourRepository extends JpaRepository<Tour, Integer> {
+public interface TourRepository extends JpaRepository<Tour, Long> {
 }
